move media-related theme classes to Media namespace
